+++ src/Kernel/include/Kernel/Settings.h
@@ -38,6 +38,7 @@
 #include <QFileInfo>
 #include <QByteArray>
 
+#include "FlightSimulatorShortcuts.h"
 #include "Replay.h"
 #include "SampleRate.h"
 #include "KernelLib.h"
@@ -502,13 +503,29 @@ public:
     QString getImportAircraftType() const noexcept;
 
     /*!
-     * Sets the aircraft type (name) for import
+     * Sets the aircraft type (name) for import.
      *
      * \param type
      *        the (last) selected aircraft type
      * \sa changed
      */
     void setImportAircraftType(const QString &type) noexcept;
+
+    /*!
+     * Returns the flight simulator shortcuts that can be triggered in the connected flight simulator.
+     *
+     * \return the flight simulator shortcuts
+     */
+    FlightSimulatorShortcuts getFlightSimulatorShortcuts() const noexcept;
+
+    /*!
+     * Sets the flight simulator shortcuts that can be triggered in the connected flight simulator.
+     *
+     * \param shortcuts
+     *        the client event shortcuts
+     * \sa flightSimulatorShortcutsChanged
+     */
+    void setFlightSimulatorShortcuts(FlightSimulatorShortcuts shortcuts) noexcept;
 
     /*!
      * Returns the file info of the best available earth gravity model (EGM) data file.
@@ -697,6 +714,13 @@ signals:
      * \sa changed
      */
     void defaultMinimalUiReplaySpeedVisibilityChanged(bool hidden);
+
+    /*!
+     * Emitted when the flight simulator shortcuts have changed.
+     *
+     * \sa changed
+     */
+    void flightSimulatorShortcutsChanged(const FlightSimulatorShortcuts &shortcuts);
 
     /*!
      * Emitted when any setting has changed.