+++ src/Kernel/src/Settings.cpp
@@ -102,8 +102,8 @@ struct SettingsPrivate
     bool defaultMinimalUiReplaySpeedVisible {DefaultMinimalUiReplaySpeedVisible};
 
     QString importAircraftType;
-
     QFileInfo earthGravityModelFileInfo;
+    FlightSimulatorShortcuts flightSimulatorShortcuts {DefaultFlightSimulatorShortcuts};
 
     int previewInfoDialogCount {DefaultPreviewInfoDialogCount};
 
@@ -138,6 +138,24 @@ struct SettingsPrivate
 
     static constexpr int DefaultPreviewInfoDialogCount {3};
     static constexpr int PreviewInfoDialogBase {150};
+
+            // TODO Select proper defaults (non-assigned by default in MSFS)
+    static inline const QKeySequence DefaultRecordShortcut {"Ctrl+R"};
+    static inline const QKeySequence DefaultReplayShortcut {"Ctrl+E"};
+    static inline const QKeySequence DefaultPauseShortcut {"Ctrl+P"};
+    static inline const QKeySequence DefaultStopShortcut {"Ctrl+S"};
+    static inline const QKeySequence DefaultBackwardShortcut {"Ctrl+B"};
+    static inline const QKeySequence DefaultForwardShortcut {"Ctrl+F"};
+    static inline const QKeySequence DefaultRewindShortcut {"Ctrl+T"};
+    static inline const FlightSimulatorShortcuts DefaultFlightSimulatorShortcuts {
+        DefaultRecordShortcut,
+        DefaultReplayShortcut,
+        DefaultPauseShortcut,
+        DefaultStopShortcut,
+        DefaultBackwardShortcut,
+        DefaultForwardShortcut,
+        DefaultRewindShortcut
+    };
 };
 
 // PUBLIC
@@ -488,6 +506,19 @@ void Settings::setImportAircraftType(const QString &type) noexcept
     }
 }
 
+FlightSimulatorShortcuts Settings::getFlightSimulatorShortcuts() const noexcept
+{
+    return d->flightSimulatorShortcuts;
+}
+
+void Settings::setFlightSimulatorShortcuts(FlightSimulatorShortcuts shortcuts) noexcept
+{
+    if (d->flightSimulatorShortcuts != shortcuts) {
+        d->flightSimulatorShortcuts = std::move(shortcuts);
+        emit flightSimulatorShortcutsChanged(d->flightSimulatorShortcuts);
+    }
+}
+
 QFileInfo Settings::getEarthGravityModelFileInfo() const noexcept
 {
     return d->earthGravityModelFileInfo;
@@ -575,11 +606,22 @@ void Settings::store() const noexcept
     d->settings.beginGroup("Plugins");
     {
         d->settings.setValue("SkyConnectPluginUuid", d->skyConnectPluginUuid);
+        d->settings.beginGroup("FlightSimulatorShortcuts");
+        {
+            d->settings.setValue("Record", d->flightSimulatorShortcuts.record);
+            d->settings.setValue("Replay", d->flightSimulatorShortcuts.replay);
+            d->settings.setValue("Pause", d->flightSimulatorShortcuts.pause);
+            d->settings.setValue("Stop", d->flightSimulatorShortcuts.stop);
+            d->settings.setValue("Backward", d->flightSimulatorShortcuts.backward);
+            d->settings.setValue("Forward", d->flightSimulatorShortcuts.forward);
+            d->settings.setValue("Rewind", d->flightSimulatorShortcuts.rewind);
+        }
+        d->settings.endGroup();
     }
     d->settings.endGroup();
     d->settings.beginGroup("Recording");
     {
-        d->settings.setValue("RecordingSampleRate", d->recordingSampleRateValue);
+        d->settings.setValue("RecordingSampleRate", d->recordingSampleRateValue);        
     }
     d->settings.endGroup();
     d->settings.beginGroup("Replay");
@@ -663,6 +705,17 @@ void Settings::restore() noexcept
     d->settings.beginGroup("Plugins");
     {
         d->skyConnectPluginUuid = d->settings.value("SkyConnectPluginUuid", d->DefaultSkyConnectPluginUuid).toUuid();
+        d->settings.beginGroup("FlightSimulatorShortcuts");
+        {
+            d->flightSimulatorShortcuts.record = d->settings.value("Record", d->DefaultRecordShortcut).toString();
+            d->flightSimulatorShortcuts.replay = d->settings.value("Replay", d->DefaultReplayShortcut).toString();
+            d->flightSimulatorShortcuts.pause = d->settings.value("Pause", d->DefaultPauseShortcut).toString();
+            d->flightSimulatorShortcuts.stop = d->settings.value("Stop", d->DefaultStopShortcut).toString();
+            d->flightSimulatorShortcuts.backward = d->settings.value("Backward", d->DefaultBackwardShortcut).toString();
+            d->flightSimulatorShortcuts.forward = d->settings.value("Forward", d->DefaultForwardShortcut).toString();
+            d->flightSimulatorShortcuts.rewind = d->settings.value("Rewind", d->DefaultRewindShortcut).toString();
+        }
+        d->settings.endGroup();
     }
     d->settings.endGroup();
     d->settings.beginGroup("Recording");
@@ -808,6 +861,8 @@ void Settings::frenchConnection() noexcept
     connect(this, &Settings::defaultMinimalUiNonEssentialButtonVisibilityChanged,
             this, &Settings::changed);
     connect(this, &Settings::defaultMinimalUiReplaySpeedVisibilityChanged,
+            this, &Settings::changed);
+    connect(this, &Settings::flightSimulatorShortcutsChanged,
             this, &Settings::changed);
 }