+++ src/PluginManager/include/PluginManager/Connect/AbstractSkyConnect.h
@@ -31,6 +31,7 @@
 #include <QObject>
 
 #include <Kernel/SampleRate.h>
+#include <Kernel/FlightSimulatorShortcuts.h>
 #include <Model/InitialPosition.h>
 #include "SkyConnectIntf.h"
 #include "Connect.h"
@@ -56,6 +57,8 @@ public:
     AbstractSkyConnect &operator=(const AbstractSkyConnect &rhs) = delete;
     AbstractSkyConnect &operator=(AbstractSkyConnect &&rhs) = delete;
     ~AbstractSkyConnect() override;
+    
+    bool setupFlightSimulatorShortcuts(FlightSimulatorShortcuts shortcuts) noexcept override;
 
     bool setUserAircraftInitialPosition(const InitialPosition &initialPosition) noexcept override;
     bool freezeUserAircraft(bool enable) const noexcept override;
@@ -78,6 +81,7 @@ public:
 
     void setPaused(Initiator initiator, bool enable) noexcept override;
     bool isPaused() const noexcept override;
+    bool isRecordingPaused() const noexcept override;
 
     void skipToBegin() noexcept override;
     void skipBackward() noexcept override;
@@ -121,7 +125,8 @@ protected:
     void createAiObjects() noexcept;
 
     virtual bool isTimerBasedRecording(SampleRate::SampleRate sampleRate) const noexcept = 0;
-
+    
+    virtual bool onSetupFlightSimulatorShortcuts(const FlightSimulatorShortcuts &shortcuts) noexcept = 0;
     virtual bool onInitialPositionSetup(const InitialPosition &initialPosition) noexcept = 0;
     virtual bool onFreezeUserAircraft(bool enable) const noexcept = 0;
     virtual bool onSimulationEvent(SimulationEvent event, float arg1) const noexcept = 0;