+++ src/Plugins/Connect/PathCreator/src/PathCreatorPlugin.h
@@ -30,6 +30,7 @@
 
 #include <QObject>
 
+#include <Kernel/FlightSimulatorShortcuts.h>
 #include <Model/TimeVariableData.h>
 #include <PluginManager/Connect/AbstractSkyConnect.h>
 #include <PluginManager/Connect/SkyConnectIntf.h>
@@ -52,11 +53,14 @@ public:
     PathCreatorPlugin &operator=(const PathCreatorPlugin &rhs) = delete;
     PathCreatorPlugin &operator=(PathCreatorPlugin &&rhs) = delete;
     ~PathCreatorPlugin() override;
+    
 
     bool setUserAircraftPosition(const PositionData &positionData) noexcept override;
 
 protected:
     bool isTimerBasedRecording(SampleRate::SampleRate sampleRate) const noexcept override;
+    
+    bool onSetupFlightSimulatorShortcuts(const FlightSimulatorShortcuts &shortcuts) noexcept override;
 
     bool onInitialPositionSetup(const InitialPosition &initialPosition) noexcept override;
     bool onFreezeUserAircraft(bool enable) const noexcept override;