+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## 0.17.0
 
+### New Features
+- Keyboard shortcuts can be defined and triggered within the flight simulator, in order to:
+  * Toggle recording / replay
+  * Pause / Stop
+  * Skip forward/backward
+  * Skip to begin
+  * The shortcuts can be defined in the Sky Dolly application settings
+- Desktop notifications indicate whether recording has started, paused/resumed and stopped
+
 ### Improvements
 - The default logbook name is now *Sky Dolly Logbook*
   * In order to help to avoid confusion between the Sky Dolly application and logbook directory
@@ -22,7 +31,7 @@
 ### Bug Fixes
 - KML export: the _first_ position point is now properly exported as well
 - Flight condition is not overwritten anymore when recording additional formation aircraft
-- Toggling the "active pause" in MSFS while recording now properly resumes/pauses recording and the flight [[Issue #106](https://github.com/till213/SkyDolly/issues/106)]
+- Toggling the "active pause" in MSFS while recording now properly resumes/pauses recording and the flight itself [[Issue #106](https://github.com/till213/SkyDolly/issues/106)]
 
 ### Under the Hood
 - Upgrade GeographicLib to version 2.2 (from version 2.1.2)